+++ scripts/build_docker.sh
@@ -13,5 +13,10 @@ else
   TAG="${TAG}-cpu"
 fi
 
-docker build --build-arg PARENT_IMAGE=${PARENT} -t ${TAG}:${VERSION} .
+docker build --build-arg PARENT_IMAGE=${PARENT} --build-arg USE_GPU=${USE_GPU} -t ${TAG}:${VERSION} .
+docker tag ${TAG}:${VERSION} ${TAG}:latest
 
+if [[ ${RELEASE} == "True" ]]; then
+  docker push ${TAG}:${VERSION}
+  docker push ${TAG}:latest
+fi