+++ stable_baselines/logger.py
@@ -233,6 +233,8 @@ class TensorBoardOutputFormat(KVWriter):
         summary = tf.Summary(value=[summary_val(k, v) for k, v in kvs.items() if valid_float_value(v)])
         event = event_pb2.Event(wall_time=time.time(), summary=summary)
         event.step = self.step  # is there any reason why you'd want to specify the step?
+        if self.writer is None:
+            raise ValueError("Attempt to write after close().")
         self.writer.WriteEvent(event)
         self.writer.Flush()
         self.step += 1